Attempt to fix savestates memory read bug
Sometimes (relatively rarely), savestates or loadstate can happen in a situation where the memory isn't in a readable state, making a panic error.
This commit tries to delay the loadstate and savestate call to the QueueHostJob, so it hopefully doesn't happen.
It also doesn't emit the event if the ppc state is incorrect, and emit a custom panic message.