Script end event
Added a script end event, which is triggered on any script end. It takes an int argument, corresponding to the id of the script which is ending.
I also added a python stub to get the id of the current script, so the user can check if the script ending is the current script or not